requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.27
python-dotenv==1.0.1
email-validator==2.1.0.post1
Pillow==10.2.0
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Session==0.5.0
Flask-Caching==2.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-Admin==1.6.1
Flask-Babel==4.0.0
Flask-DebugToolbar==0.14.1
Flask-Principal==0.4.0
Flask-Security==5.6.1
Flask-Uploads==0.2.1
Flask-User==1.0.2.2
Flask-APScheduler==1.13.1
Flask-Assets==2.1.0
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-Compress==1.14
Flask-Dropzone==2.0.0
Flask-Excel==0.0.7
Flask-Gravatar==0.5.0
Flask-HTTPAuth==4.8.0
Flask-Images==0.1.0
Flask-JSON==0.4.0
Flask-JSONRPC==0.3.1
Flask-Limiter==3.5.0
Flask-Marshmallow==0.15.0
Flask-Moment==1.0.4
Flask-OAuthlib==0.9.6
Flask-PageDown==0.2.2
Flask-Paginate==2022.1.8
Flask-PyMongo==2.3.0
Flask-RESTX==1.3.0
Flask-Script==2.0.6
Flask-Sijax==0.4.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Static-Digest==0.1.4
Flask-Testing==0.8.1
Flask-Themes2==0.1.5
Flask-Uploads==0.2.1
Flask-WhooshAlchemy==0.56
Flask-WhooshAlchemy3==0.7.6
Flask-WTF==1.2.1
Flask-XML-RPC==0.1.1
Flask-Zipkin==0.1.0
pycryptodome==3.20.0
mysql-connector-python==8.3.0
